# book-network-ui/src/app/services/models/book-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BookRequest {
  authorName: string;
  id?: number;
  isbn: string;
  sharable?: boolean;
  synopsis: string;
  title: string;
}