# book-network-ui/src/app/services/models/book-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BookResponse {
  archived?: boolean;
  authorName?: string;
  cover?: Array<string>;
  id?: number;
  isbn?: string;
  owner?: string;
  rate?: number;
  sharable?: boolean;
  synopsis?: string;
  title?: string;
}